+++ src/main/java/com/luxoft/wheretogo/services/UsersService.java
@@ -10,7 +10,7 @@ public interface UsersService {
 
 	List<User> findAll();
 
-	User findById(int userId);
+	User findById(long userId);
 
 	User findByLogin(String userLogin);
